Update Dockerfile and app.py to include new trained model
The Dockerfile and app.py have been updated to include a new trained model. This model, saved_model999.zip, is now being copied into the Docker container and used by the application if the expected model is not found or fails to download. The exposed port number has also been updated to port 52525.